requirements.txt
fastapi>=0.115.0
httpx>=0.27.0
pydantic-ai>=0.0.39
python-dotenv>=1.0.0
sse-starlette>=2.0.0
uvicorn[standard]>=0.30.0
